Remove `tcpKeepAliveListener`
From Go 1.13, this code was removed from the Go source. Instead, a
`net.TCPListener` has a configurable `KeepAlive`, which defaults to
`net.defaultTCPKeepAlive`. This is currently set to 15 seconds, as of Go
1.15. (It was originally set as 3 minutes, because the code was
copy-pasted from Go's source.)

This change simplifies our code and changes our server behaviour to be
consistent with the default in Go, which is also the behaviour in the
calls to `http.Server.ListenAndServe()` that we use in main.go.

See golang/go#23378 and golang/go#31510.